+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## [1.0.8-beta.3](https://github.com/avohq/react-native-analytics-debugger/compare/1.0.8-beta.2...1.0.8-beta.3) (2019-09-10)
+
+
+### Bug Fixes
+
+* Fix reason type def ([bf0494e](https://github.com/avohq/react-native-analytics-debugger/commit/bf0494e))
+
+
+
+## [1.0.8-beta.2](https://github.com/avohq/react-native-analytics-debugger/compare/1.0.8-beta.1...1.0.8-beta.2) (2019-09-10)
+
+
+### Bug Fixes
+
+* Pass mode strings correctly from reason bindings ([077113e](https://github.com/avohq/react-native-analytics-debugger/commit/077113e))
+
+
+
+## [1.0.8-beta.1](https://github.com/avohq/react-native-analytics-debugger/compare/1.0.8-beta.0...1.0.8-beta.1) (2019-09-10)
+
+
+### Code Refactoring
+
+* Make the API more explicit ([6da0018](https://github.com/avohq/react-native-analytics-debugger/commit/6da0018))
+
+
+### BREAKING CHANGES
+
+* enable/disable are no longer supported, use showDebugger/hideDebugger
+
+
+
+## [1.0.8-beta.0](https://github.com/avohq/react-native-analytics-debugger/compare/1.0.7...1.0.8-beta.0) (2019-09-10)
+
+
+
 ## [1.0.7](https://github.com/avohq/react-native-analytics-debugger/compare/1.0.7-beta.0...1.0.7) (2019-09-10)